WA: https://atcoder.jp/contests/panasonic2020/submissions/11061339

+++ panasonic2020/c.py
@@ -1,5 +1,5 @@
 #!/usr/bin/python3
-a, b, c = map(float, input().split())
+a, b, c = map(int, input().split())
 if (a*b)*4 < ((c-a-b))**2:
     print("Yes")
 else:
